Move ActorRegistry back to environment.rs
ActorRegistry belongs to the build phase (environment), not the
scheduler. The scheduler consumes it via RunHandle but does not own it.
