+++ ailets-rs/ailetos/src/lib.rs
@@ -43,12 +43,12 @@ pub use fd_table::{FdEntry, FdTable};
 pub use stub_actor_runtime::{BlockingActorRuntime, ShutdownHandle};
 
 // Re-export environment types
-pub use environment::{Environment, RunHandle};
+pub use environment::{ActorRegistry, Environment, RunHandle};
 
 // Re-export suspension types
 pub use suspension::SuspensionState;
 
 // Re-export scheduler
 pub use scheduler::{
-    is_ready_to_spawn, run_spawn_loop, ActorFn, ActorRegistry, Scheduler, StopConditions,
+    is_ready_to_spawn, run_spawn_loop, ActorFn, Scheduler, StopConditions,
 };
